src/requirements.txt
# frozen from registry

crunch-cli
networkx
numpy
pandas
pytorch-lightning
scikit-learn==1.8.0
torch==2.10.0
tqdm
